+++ ntex-util/src/future/join.rs
@@ -105,7 +105,7 @@ mod tests {
     use super::*;
     use crate::{future::Ready, time};
 
-    #[ntex_macros::rt_test2]
+    #[ntex::test]
     async fn join_tests() {
         let res = join(Ready::<_, ()>::Ok("test"), time::sleep(time::Millis(50))).await;
         assert_eq!(res, (Ok("test"), ()));